pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "charliehustle"
version = "2.0.0"
description = "MLB game prediction and betting simulation"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Taylor Seale" },
]
dependencies = [
    "MLB-StatsAPI>=1.7",
    "pybaseball>=2.3",
    "xgboost>=2.0",
    "scikit-learn>=1.4",
    "pandas>=2.1",
    "numpy>=1.26",
    "matplotlib>=3.8",
    "click>=8.1",
    "pydantic>=2.5",
    "joblib>=1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
charliehustle = "charliehustle.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/charliehustle"]

[tool.pytest.ini_options]
testpaths = ["tests"]
